feat(core-contract): relocate aiConfigManifest + coreOperationalManifest to ./settings (PRD-239 US-01)
Moves the two core-pillar settings manifests out of
packages/module-registry/src/settings/core/ into
packages/core-contract/src/settings/, exposes a ./settings subpath on
@pops/core-contract, and repoints @pops/pillar-sdk/settings at the new
package for those two named exports. Pure relocation -- SettingsManifest
shape, export names, and group contents are unchanged.

- adds packages/core-contract/src/settings/{ai-manifest.ts,operational-manifest.ts,index.ts}
- declares "./settings" subpath in core-contract/package.json exports
- pillar-sdk re-exports aiConfigManifest + coreOperationalManifest from
  @pops/core-contract/settings (other eight manifests still flow through
  @pops/module-registry/settings pending US-02/03/04/05)
- promotes @pops/core-contract to a runtime dep on pillar-sdk (the barrel
  now imports it at runtime)
- drops the two re-export lines from module-registry's settings barrel
  and deletes packages/module-registry/src/settings/core/
- repoints module-registry's internal known-modules.ts script at
  @pops/core-contract/settings (script is retired with PRD-218 US-03)
- ai-manifest.test.ts moves to core-contract/src/__tests__/ alongside the
  package's existing test layout
- smoke test in core-contract asserts aiConfigManifest.id === 'ai.config'
  and coreOperationalManifest.id === 'core.operational'

Stubs CoreRouter as AnyTRPCRouter to break the
pillar-sdk -> core-contract -> core-api -> pillar-sdk build-graph cycle
that opening pillar-sdk's runtime dep on core-contract would close.
Mirrors the precedent established for @pops/finance-contract in #2998.
PRD-155's declaration bundler will restore concrete per-procedure types
without re-introducing the cycle.

Refs PRD-239 (#3204), PR #3175, PR #3176.
