+++ packages/core-contract/src/settings/ai-manifest.ts
