+++ packages/core-contract/src/settings/index.ts
@@ -1,0 +1,2 @@
+export { aiConfigManifest } from './ai-manifest.js';
+export { coreOperationalManifest } from './operational-manifest.js';
