+++ packages/core-contract/src/settings/operational-manifest.ts
