+++ packages/module-registry/src/settings/index.ts
@@ -4,8 +4,6 @@
  * (`apps/pops-api`) read from a single source of truth. Pure data, no
  * runtime dependencies beyond `@pops/types`.
  */
-export { aiConfigManifest } from './core/ai-manifest.js';
-export { coreOperationalManifest } from './core/operational-manifest.js';
 export { financeManifest } from './finance/index.js';
 export { inventoryManifest } from './inventory/index.js';
 export { cerebrumManifest } from './cerebrum/index.js';
