+++ packages/pillar-sdk/src/settings/index.ts
@@ -1,12 +1,13 @@
 /**
- * Pillar-scoped settings manifests, re-exported from per-pillar contract
- * packages so consumers can pull settings off the SDK surface (PRD-238 US-01,
- * Option B). Media now lives in `@pops/media-contract/settings` (PRD-239
- * US-05); the remaining manifests still ride on `@pops/module-registry/settings`
- * until they are relocated.
+ * Pillar-scoped settings manifests, re-exported from each pillar's own
+ * contract package as the relocation in PRD-239 progresses. Manifests not yet
+ * relocated still flow through `@pops/module-registry/settings`; once PRD-239
+ * US-01..US-05 land, this barrel reads exclusively from the per-pillar
+ * packages and the module-registry subpath is retired by US-06.
  *
  * Pure re-export — no shape change, no runtime behaviour change.
  */
+export { aiConfigManifest, coreOperationalManifest } from '@pops/core-contract/settings';
 export {
   arrManifest,
   plexManifest,
@@ -14,8 +15,6 @@ export {
   mediaOperationalManifest,
 } from '@pops/media-contract/settings';
 export {
-  aiConfigManifest,
-  coreOperationalManifest,
   inventoryManifest,
   financeManifest,
   cerebrumManifest,
